Fix jest test in CI
Jest runs out of memory when run in Docker, possibly because of a memory
leak in ts-jest.

The tests will now run after the build, on the built *.js files. As a
consequence, the jest tests will run faster because ts-jest will no
longer be invoked in ci.

Check the package.json test:ci script for more info.

https://github.com/facebook/jest/issues/7874
https://github.com/facebook/jest/issues/9081